added finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,11 +1,11 @@
-the_income = int(input("Enter your monthly income:"))
-total_expenses = int(input ("Enter your total monthly expenses:"))
+the_income = float(input("Enter your monthly income:"))
+total_expenses = float(input ("Enter your total monthly expenses:"))
 
-monthly_savings = total_expenses - the_income 
+monthly_savings = float(total_expenses - the_income)
 
 #(Projected Savings = Monthly Savings * 12 + (Monthly Savings * 12 * 0.05)).
 
-Projected_Savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+Projected_Savings = float(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print(f"""
 Your monthly savings are ${monthly_savings}.
